LayoutTests/svg/animations/animate-keySplines-expected.txt

SVG 1.1 dynamic animation tests

Testing correct parsing of keySplines.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.height.animVal.value is 167
PASS rect.height.baseVal.value is 100
PASS rect.height.animVal.value is 111
PASS rect.height.baseVal.value is 100
PASS rect.height.animVal.value is 0
PASS rect.height.baseVal.value is 100
PASS rect.height.animVal.value is 100
PASS rect.height.baseVal.value is 100
PASS successfullyParsed is true

TEST COMPLETE
